requirements.txt
numpy<2.0.0
Pillow>=9.0.0
matplotlib>=3.4.0
